task3 add DBService.java

+++ Simple Banking System/task/src/banking/Account.java
@@ -3,11 +3,11 @@ package banking;
 public class Account {
     String number;
     String pin;
-    long balance;
+    int balance;
 
     public Account(String number, String pin) {
         this.number = number;
         this.pin = pin;
-        this.balance = 0L;
+        this.balance = 0;
     }
 }
